+++ src/modules/tools/fileImport/store/actionsFileImport.js
@@ -140,6 +140,31 @@ function checkIsVisibleSetting (features) {
     return resFeatures;
 }
 
+/**
+ * Gets custom attributes from feature by parsing given feature keys.
+ * @param {ol/Feature} feature The feature.
+ * @returns {Object} The parsed attributes.
+ */
+function getParsedCustomAttributes (feature) {
+    if (typeof feature.getKeys !== "function" || !Array.isArray(feature.getKeys())) {
+        return {};
+    }
+
+    let attributes = feature.get("attributes");
+
+    if (!isObject(attributes)) {
+        attributes = {};
+    }
+    feature.getKeys().forEach(key => {
+        if (typeof key.split === "function" && key.split("custom-attribute____").length > 1) {
+            const parsedKey = key.split("custom-attribute____")[1];
+
+            attributes[parsedKey] = feature.get(key);
+        }
+    });
+    return attributes;
+}
+
 export default {
     /**
      * Sets the selected file type
@@ -155,18 +180,19 @@ export default {
      * Sets the featureExtents
      * @param {Object} param.state the state
      * @param {Object} param.commit the commit
-     * @param {ol/Feature[]} features the parsed features
+     * @param {ol/Feature[]} payload.features the parsed features
+     * @param {String} payload.fileName the file name
      * @returns {void}
      */
-    setFeatureExtents: ({state, commit}, features) => {
-        const extents = [... state.featureExtents],
+    setFeatureExtents: ({state, commit}, {features, fileName}) => {
+        const extents = state.featureExtents,
             extent = createEmptyExtent();
 
         for (let i = 0; i < features.length; i++) {
             extend(extent, features[i].getGeometry().getExtent());
         }
 
-        extents.push(extent);
+        extents[fileName] = extent;
         commit("setFeatureExtents", extents);
     },
 
@@ -181,8 +207,10 @@ export default {
     importKML: ({state, dispatch, rootGetters}, datasrc) => {
         const
             vectorLayer = datasrc.layer,
-            format = getFormat(datasrc.filename, state.selectedFiletype, state.supportedFiletypes, supportedFormats),
-            crsPropName = getCrsPropertyName(datasrc.raw);
+            fileName = datasrc.filename,
+            format = getFormat(fileName, state.selectedFiletype, state.supportedFiletypes, supportedFormats),
+            crsPropName = getCrsPropertyName(datasrc.raw),
+            customAttributes = {};
 
         let
             featureError = false,
@@ -194,7 +222,7 @@ export default {
         }
 
         if (format === false) {
-            const fileNameSplit = datasrc.filename.split("."),
+            const fileNameSplit = fileName.split("."),
                 fileFormat = fileNameSplit.length > 0 ? "*." + fileNameSplit[fileNameSplit.length - 1] : "unknown";
 
             alertingMessage = {
@@ -239,7 +267,7 @@ export default {
             console.warn(ex);
             alertingMessage = {
                 category: i18next.t("common:modules.alerting.categories.error"),
-                content: i18next.t("common:modules.tools.fileImport.alertingMessages.formatError", {filename: datasrc.filename})
+                content: i18next.t("common:modules.tools.fileImport.alertingMessages.formatError", {filename: fileName})
             };
 
             dispatch("Alerting/addSingleAlert", alertingMessage, {root: true});
@@ -249,7 +277,7 @@ export default {
         if (!Array.isArray(features) || features.length === 0) {
             alertingMessage = {
                 category: i18next.t("common:modules.alerting.categories.error"),
-                content: i18next.t("common:modules.tools.fileImport.alertingMessages.missingFileContent", {filename: datasrc.filename})
+                content: i18next.t("common:modules.tools.fileImport.alertingMessages.missingFileContent", {filename: fileName})
             };
 
             dispatch("Alerting/addSingleAlert", alertingMessage, {root: true});
@@ -257,7 +285,16 @@ export default {
         }
 
         features.forEach(feature => {
+            const featureAttributes = getParsedCustomAttributes(feature);
             let geometries;
+
+            feature.set("attributes", featureAttributes);
+            feature.setProperties(featureAttributes);
+            Object.keys(featureAttributes).forEach(key => {
+                if (!Object.prototype.hasOwnProperty.call(customAttributes, key)) {
+                    customAttributes[key] = key;
+                }
+            });
 
             if (feature.get("isGeoCircle")) {
                 const circleCenter = feature.get("geoCircleCenter").split(",").map(parseFloat),
@@ -300,31 +337,33 @@ export default {
                     }
 
                     geometry.transform(mappedCrsPropName, rootGetters["Maps/projectionCode"]);
+                    feature.set("source", fileName);
                 });
             }
         });
         features = checkIsVisibleSetting(features);
 
         vectorLayer.getSource().addFeatures(features);
+        vectorLayer.set("gfiAttributes", customAttributes);
 
         if (featureError) {
             alertingMessage = {
                 category: i18next.t("common:modules.alerting.categories.info"),
-                content: i18next.t("common:modules.tools.fileImport.alertingMessages.successPartly", {filename: datasrc.filename})
+                content: i18next.t("common:modules.tools.fileImport.alertingMessages.successPartly", {filename: fileName})
             };
         }
         else {
             alertingMessage = {
                 category: i18next.t("common:modules.alerting.categories.info"),
-                content: i18next.t("common:modules.tools.fileImport.alertingMessages.success", {filename: datasrc.filename})
+                content: i18next.t("common:modules.tools.fileImport.alertingMessages.success", {filename: fileName})
             };
         }
 
         dispatch("Alerting/addSingleAlert", alertingMessage, {root: true});
-        dispatch("addImportedFilename", datasrc.filename);
+        dispatch("addImportedFilename", fileName);
 
         if (state.enableZoomToExtend && features.length) {
-            dispatch("setFeatureExtents", features);
+            dispatch("setFeatureExtents", {features: features, fileName: fileName});
         }
     },
 
@@ -338,7 +377,8 @@ export default {
      */
     importGeoJSON: ({state, dispatch, rootGetters}, datasrc) => {
         const vectorLayer = datasrc.layer,
-            format = getFormat(datasrc.filename, state.selectedFiletype, state.supportedFiletypes, supportedFormats),
+            fileName = datasrc.filename,
+            format = getFormat(fileName, state.selectedFiletype, state.supportedFiletypes, supportedFormats),
             gfiAttributes = {};
 
         let
@@ -346,7 +386,7 @@ export default {
             features;
 
         if (format === false) {
-            const fileNameSplit = datasrc.filename.split("."),
+            const fileNameSplit = fileName.split("."),
                 fileFormat = fileNameSplit.length > 0 ? "*." + fileNameSplit[fileNameSplit.length - 1] : "unknown";
 
             alertingMessage = {
@@ -365,7 +405,7 @@ export default {
             console.warn(ex);
             alertingMessage = {
                 category: i18next.t("common:modules.alerting.categories.error"),
-                content: i18next.t("common:modules.tools.fileImport.alertingMessages.formatError", {filename: datasrc.filename})
+                content: i18next.t("common:modules.tools.fileImport.alertingMessages.formatError", {filename: fileName})
             };
 
             dispatch("Alerting/addSingleAlert", alertingMessage, {root: true});
@@ -375,7 +415,7 @@ export default {
         if (!Array.isArray(features) || features.length === 0) {
             alertingMessage = {
                 category: i18next.t("common:modules.alerting.categories.error"),
-                content: i18next.t("common:modules.tools.fileImport.alertingMessages.missingFileContent", {filename: datasrc.filename})
+                content: i18next.t("common:modules.tools.fileImport.alertingMessages.missingFileContent", {filename: fileName})
             };
 
             dispatch("Alerting/addSingleAlert", alertingMessage, {root: true});
@@ -526,6 +566,7 @@ export default {
                 geometry.transform("EPSG:4326", rootGetters["Maps/projectionCode"]);
             });
 
+            feature.set("source", fileName);
             vectorLayer.getSource().addFeature(feature);
         });
 
@@ -535,14 +576,14 @@ export default {
 
         alertingMessage = {
             category: i18next.t("common:modules.alerting.categories.info"),
-            content: i18next.t("common:modules.tools.fileImport.alertingMessages.success", {filename: datasrc.filename})
+            content: i18next.t("common:modules.tools.fileImport.alertingMessages.success", {filename: fileName})
         };
 
         dispatch("Alerting/addSingleAlert", alertingMessage, {root: true});
-        dispatch("addImportedFilename", datasrc.filename);
+        dispatch("addImportedFilename", fileName);
 
         if (state.enableZoomToExtend && features.length) {
-            dispatch("setFeatureExtents", features);
+            dispatch("setFeatureExtents", {features: features, fileName: fileName});
         }
     },
     /**